backend/pyproject.toml
 [project]
 name = "countit-backend"
 version = "0.1.0"
 description = "countIT backend (FastAPI + Postgres + auth + products)"
 requires-python = ">=3.11"
 dependencies = [
   "fastapi>=0.115.0",
   "uvicorn[standard]>=0.30.0",
   "pydantic-settings>=2.3.0",
  "email-validator>=2.0.0",
   "sqlalchemy>=2.0.30",
   "alembic>=1.13.2",
   "psycopg[binary]>=3.2.0",
   "httpx>=0.27.0",
   "passlib[bcrypt]>=1.7.4",
   "python-jose[cryptography]>=3.3.0",
 ]
 
 [tool.uv]
 dev-dependencies = [
   "ruff>=0.6.0",
 ]
 
